SafePill_deploy_ready/SafePill_deploy_ready/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
numpy==1.26.4
Pillow==11.1.0
requests==2.32.3
openpyxl==3.1.5
pandas==2.2.3
easyocr==1.7.2
